fix(cron): satisfy configured auth CI contracts
